Update Problem Markdown Generator

+++ tools/build.sh
@@ -2,18 +2,25 @@
 
 echo '清除旧目录'
 rm -rf SUMMARY.md
-rm -rf tools/tpl
 rm -rf tools/markdowns
 
 echo '创建目录'
-mkdir tools/tpl
 mkdir tools/markdowns
+
+if [ $1 ]
+then
+    echo '单独指定生成 Problem '$1'，跳过清空全部 Problem Markdown 操作'
+else
+    echo '执行清空全部 Problem List 操作'
+    rm -rf tools/tpl
+    mkdir tools/tpl
+fi
 
 echo '准备构建 SUMMARY'
 node tools/createSummary.js
 
 echo '准备构建模板'
-node tools/createMarkdown.js
+node tools/createMarkdown.js $1
 
 echo '准备安装依赖'
 cd . && gitbook install
@@ -21,10 +28,5 @@ cd . && gitbook install
 echo '输出文件'
 cd . && gitbook build
 
-if [ $1 ]
-then
-    echo '服务器脚本不执行 gitbook server'
-else 
-    echo '开始打包 gitbook'
-    cd . && gitbook serve
-fi
+echo '开始打包 gitbook'
+cd . && gitbook serve
